requirements.txt
# requirements.txt

streamlit==1.22.0
numpy==1.23.5
pandas==2.0.1
feat==1.0.5
keras==2.12.0
matplotlib==3.7.1
opencv-python==4.7.0.72
Pillow==9.5.0
requests==2.29.0
torchvision==0.15.1
